Reformat MainForm files, add margin to the capture menu
Sorry for the big diff, the designer automatically reformats the file on save,
and there's no way to disable that.
About the capture menu, I tried my best to put all the controls inside one
container to make them flow properly, and I failed (not really a WinForms
developer, and this bad experience will not convince me to be one), so I just
added some more margin to the menu and leave it at that, assuming that the UI
will be rewritten at some point.